+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Updated screenshots.
+- Update to latest from beautiful-jekyll.
+
+### Added
+
+- Adds whitespace check exclusion for problematic beautiful-jekyll sources.
 
 ## [v3.23.0](https://github.com/lexicalunit/spellbot/releases/tag/v3.23.0) - 2020-08-25